requirements.txt
pandas
scikit-learn
seaborn
matplotlib
opencv-python-headless
tensorflow
dvc
pyyaml
python-box
ensure
gdown
keras
Flask
Wekzeug
numpy
tqdm
joblib
